Weapon Hotbar, Modular Weapon Scripting, Misc. Weapon Updates
Added a non-persistent physical weapon inventory which can be toggled on and off using the menu button (for now). Weapons can be added to the hotbar by dropping them into an empty slot. Began converting weapon scripts to a component-based/modular design; all required behaviours for reloading are now located in the Reloadable component. Added PWM control to railgun haptic feedback. Fixed a null exception in the flamethrower script that would cause it to continue firing when dropped.