Improved Reload Mechanics
Magazines no longer instantly snap into place when reloading. When a compatible magazine is detected at the opening of the magwell, the magazine is automatically aligned with the magwell chamber and magazine movement is constrained to slide along the central axis of the magwell. Sliding the magazine to the top of the magwell reloads as before.