Shield Functionality, Hover Highlighting, Fabricator List Update
Added a script to control shield health with recharge and modified TakeDamage method to display damage text in cyan. Added a script to highlight interactable items when the controller is in their hitbox. Updated the Fabricator's item list with all missing weapons except LR, which still needs a thumbnail. Modified the Fabricator "Buy" button to stop accidentally spawning multiple items in one push.