Armory, Tidying Up Shop Scripts
Added a prototype of the Armory. Displays owned weapons as a grid of buttons (will be scrollable). Currently, pressing a button spawns that weapon to the right using the dissolve shader and updates the armory count (will change this in future to open a details/stats screen with buttons to withdraw the weapon or return to the main screen). Dropping a weapon on the table to the left destroys it and updates the inventory count. Still to do: Dissolve-out deposited weapons; debounce buttons; add scripting to "next" and "previous" buttons; create and script details pane. Moved the additional code added to the Fabricator to a separate script and tidied up the implementation. Still to do: write a method to extract weapon stats; add readouts for resource costs and resources owned; fix Start() nulls.