Loot Prefabs, Highlight Improvements, Optimisation
Added/completed missing and unfinished loot item prefabs. All are tested and working. Reworked the HoverHighlight script to prevent errors caused by destruction of held items. Objects and components deactivated in the inspector which are initialised on Start() will initialise when activated in-game (potential framerate impact). Began process of eliminating this behaviour by either initalising on Awake() or activating objects/components in the inspector and deactivating on completion of Start().