Tutorial Fixes, Black Box/Loot Inventory Fix, Healthbar Fix
Modified waypoint locations in the Tutorial to work with the rescaled prefabs & tidied up the scene hierarchy. Fixed an issue preventing the Black Box from being added to the loot inventory when the Black Box object is not at the root of the transform. Fixed an issue with the HealthBar script where targets could continue to take damage after death, repeatedly calling die().