Fixed Shield Bugs, Added Airlock to Dungeon, PlayerState Update
Enemy state changes previously only deactivated the visual effects of the shield. Added a method to ShieldBar to correctly deactivate the shield behaviour, allowing direct damage to enemy health when in Patrol mode. Added the missing airlock components to the prefab used for generation and updated scripts for compatibility with the dungeon generation process. Modified SceneLoadTest to clear the player's inCombat status when changing scenes, including on player death.